Add support for Visible History to timeseries space views
