+++ crates/re_data_store/src/entity_properties.rs
@@ -237,6 +237,11 @@ impl VisibleHistory {
         to: VisibleHistoryBoundary::AT_CURSOR,
     };
 
+    pub const ALL: Self = Self {
+        from: VisibleHistoryBoundary::Infinite,
+        to: VisibleHistoryBoundary::Infinite,
+    };
+
     pub fn from(&self, cursor: TimeInt) -> TimeInt {
         match self.from {
             VisibleHistoryBoundary::Absolute(value) => TimeInt::from(value),
