+++ crates/re_space_view_time_series/Cargo.toml
@@ -17,15 +17,16 @@ all-features = true
 
 [dependencies]
 re_arrow_store.workspace = true
+re_data_ui.workspace = true
+re_format.workspace = true
 re_log_types.workspace = true
+re_query.workspace = true
 re_renderer.workspace = true
 re_space_view.workspace = true
 re_tracing.workspace = true
 re_types.workspace = true
 re_ui.workspace = true
 re_viewer_context.workspace = true
-re_query.workspace = true
-re_format.workspace = true
 
 egui_plot.workspace = true
 egui.workspace = true
